+++ src/stores/Call.ts
@@ -12,6 +12,8 @@ import { getPartyName } from '../stores/contactStore'
 import { checkPermForCall } from '../utils/permissions'
 import { BrekekeUtils } from '../utils/RnNativeModules'
 import { waitTimeout } from '../utils/waitTimeout'
+import { authPBX } from './AuthPBX'
+import { getAuthStore } from './authStore'
 import type { CallStore } from './callStore2'
 import { contactStore } from './contactStore'
 import { intlDebug } from './intl'
@@ -195,6 +197,8 @@ export class Call {
         : intlDebug`Failed to start recording the call`
       RnAlert.error({ message, err })
     }
+    // try to re-auth if error is timeout
+    this.checkTimeoutToReconnectPbx(err)
   }
 
   toggleHoldWithCheck = () => {
@@ -213,14 +217,34 @@ export class Call {
     if (!this.isAboutToHangup && fn === 'unhold') {
       this.store.setCurrentCallId(this.id)
     }
+
     return pbx[`${fn}Talker`](this.pbxTenant, this.pbxTalkerId)
       .then(this.onToggleHoldFailure)
       .catch(this.onToggleHoldFailure)
+  }
+
+  private checkTimeoutToReconnectPbx = (err: Error | boolean) => {
+    if (err === true) {
+      return
+    }
+    if (
+      err &&
+      typeof err === 'object' &&
+      (('code' in err && (err as any).code === -1) ||
+        ('message' in err && /timeout/i.test(err.message)))
+    ) {
+      getAuthStore().pbxState = 'stopped'
+      authPBX.dispose()
+      authPBX.auth()
+    }
   }
   @action private onToggleHoldFailure = (err: Error | boolean) => {
     if (err === true) {
       return true
     }
+    // try to re-auth if error is timeout
+    this.checkTimeoutToReconnectPbx(err)
+
     const prevFn = this.holding ? 'hold' : 'unhold'
     this.setHolding(prevFn === 'unhold')
     if (typeof err !== 'boolean') {
@@ -269,6 +293,8 @@ export class Call {
       message: intlDebug`Failed to transfer the call`,
       err,
     })
+    // try to re-auth if error is timeout
+    this.checkTimeoutToReconnectPbx(err)
   }
 
   @action stopTransferring = () => {
@@ -287,6 +313,8 @@ export class Call {
       message: intlDebug`Failed to stop the transfer`,
       err,
     })
+    // try to re-auth if error is timeout
+    this.checkTimeoutToReconnectPbx(err)
   }
 
   @action conferenceTransferring = () => {
@@ -305,6 +333,8 @@ export class Call {
       message: intlDebug`Failed to make conference for the transfer`,
       err,
     })
+    // try to re-auth if error is timeout
+    this.checkTimeoutToReconnectPbx(err)
   }
 
   @action park = (number: string) =>
@@ -316,6 +346,8 @@ export class Call {
       message: intlDebug`Failed to park the call`,
       err,
     })
+    // try to re-auth if error is timeout
+    this.checkTimeoutToReconnectPbx(err)
   }
 
   private _autorunEmitEmbed = false // check if autorun is already started